Centralized styles and theme commit 3
